♻️ refactor: move RecentItem to shared types
