+++ packages/types/src/index.ts
@@ -39,6 +39,7 @@ export * from './notification';
 export * from './plugins';
 export * from './project';
 export * from './rag';
+export * from './recent';
 export * from './redis';
 export * from './resourceTransfer';
 export * from './search';
