+++ src/services/recent/index.ts
@@ -1,5 +1,6 @@
+import type { RecentItem } from '@lobechat/types';
+
 import { lambdaClient } from '@/libs/trpc/client';
-import { type RecentItem } from '@/server/routers/lambda/recent';
 
 export const RECENT_SIDEBAR_TYPES = [
   'document',
